samples/python/agents/a2a_mcp/pyproject.toml
[project]
name = "a2a-mcp"
version = "0.1.0"
description = "A2A + MCP Sample"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "fastmcp>=2.2.2",
    "google-adk>=0.3.0",
    "google-cloud-aiplatform>=1.90.0",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.6.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "a2a-samples",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "serpapi>=0.1.5",
    "langchain-mcp-adapters>=0.0.11",
]

[project.scripts]
a2a-mcp = "a2a_mcp:main"

[tool.uv.sources]
a2a-samples = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.11.7"]

[tool.ruff]
extend-select = ["I"]
